src/pyproject.toml
[project]
name = "oled-blackmeter"
version = "0.1.0"
description = "Telegram Bot for calculating the OLED black pixels percentage from images"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "anyio==4.4.0",
    "arpeggio==2.0.2",
    "attrs==24.3.0",
    "black==24.4.2",
    "blinker==1.8.2",
    "certifi==2024.7.4",
    "cfgv==3.4.0",
    "click==8.1.7",
    "cloudevents==1.10.1",
    "debugpy==1.8.12",
    "deprecation==2.1.0",
    "dill==0.3.9",
    "distlib==0.3.8",
    "filelock==3.14.0",
    "flask==3.0.3",
    "flask-wtf==1.2.1",
    "functions-framework==3.7.0",
    "grpcio==1.66.2",
    "gunicorn==22.0.0",
    "h11==0.14.0",
    "httpcore==1.0.5",
    "httpx==0.27.0",
    "identify==2.5.36",
    "idna==3.7",
    "itsdangerous==2.2.0",
    "jinja2==3.1.5",
    "markupsafe==2.1.5",
    "mypy-extensions==1.0.0",
    "nodeenv==1.9.1",
    "packaging==24.0",
    "parver==0.5",
    "pathspec==0.12.1",
    "pillow==11.1.0",
    "pip==24.3.1",
    "platformdirs==4.2.2",
    "pre-commit==3.7.1",
    "protobuf==4.25.5",
    "pulumi==3.146.0",
    "pulumi-gcp==8.14.0",
    "python-dotenv==1.0.1",
    "python-telegram-bot==21.3",
    "pyyaml==6.0.2",
    "semver==2.13.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "virtualenv==20.26.6",
    "watchdog==4.0.1",
    "werkzeug==3.0.6",
    "wtforms==3.1.2",
]

[dependency-groups]
dev = [
    "numpy>=1.26.4",
]
